+++ Sparks/Levels/LevelActions.cs
@@ -6,8 +6,8 @@
     not use this file except in compliance with the Licenses. You may
     obtain a copy of the Licenses at
     
-    https://opensource.org/license/ecl-2-0/
-    https://www.gnu.org/licenses/gpl-3.0.html
+    http://www.opensource.org/licenses/ecl2.php
+    http://www.gnu.org/licenses/gpl-3.0.html
     
     Unless required by applicable law or agreed to in writing,
     software distributed under the Licenses are distributed on an "AS IS"
@@ -18,6 +18,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using GoldenSparks.Blocks;
 using GoldenSparks.Blocks.Extended;
 using GoldenSparks.Bots;
 using GoldenSparks.DB;
@@ -26,10 +27,10 @@ using GoldenSparks.Levels.IO;
 using GoldenSparks.SQL;
 using GoldenSparks.Util;
 
-namespace GoldenSparks 
-{    
-    public static class LevelActions 
-    {       
+namespace GoldenSparks {
+    
+    public static class LevelActions {
+        
         static string BlockPropsLvlPath(string map) { return Paths.BlockPropsPath("_" + map); }
         static string BlockPropsOldPath(string map) { return Paths.BlockPropsPath("lvl_" + map); }
         
@@ -175,19 +176,24 @@ namespace GoldenSparks
         }
         
         static void DeleteDatabaseTables(string map) {
-            Database.DeleteTable("Block" + map);
+            if (Database.TableExists("Block" + map)) {
+                Database.DeleteTable("Block" + map);
+            }
             
             object locker = ThreadSafeCache.DBCache.GetLocker(map);
             lock (locker) {
                 Portal.DeleteAll(map);
                 MessageBlock.DeleteAll(map);
-                Database.DeleteTable("Zone" + map);
+                
+                if (Database.TableExists("Zone" + map)) {
+                    Database.DeleteTable("Zone" + map);
+                }
             }
         }
         
         
         public static void Replace(Level old, Level lvl) {
-            old.SaveBlockDBChanges();
+            LevelDB.SaveBlockDB(old);
             LevelInfo.Remove(old);
             LevelInfo.Add(lvl);
             
@@ -368,6 +374,7 @@ namespace GoldenSparks
             lvl.IsMuseum = true;
             
             Level.LoadMetadata(lvl);
+
             lvl.BuildAccess.Min = LevelPermission.Sparkie;
             lvl.Config.Physics = 0;
             return lvl;
@@ -393,7 +400,7 @@ namespace GoldenSparks
                 return clone;
             }
             
-            return IMapImporter.Decode(path, name, false);
+            return IMapImporter.Read(path, name, false);
         }
